requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Flask>=2.3.0
Pillow>=9.5.0
pytest>=7.4.0
